Separated out the ValidationTests and MessagingCenterTests into different files. Finished creating tests for MessagingCenter.
